# docs/openapi-chatgpt.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyMoney Expense Tracker API for ChatGPT Custom Actions",
    "description": "Secure API endpoints allowing ChatGPT to manage expenses, budgets, savings goals, and spending analytics for the authenticated MyMoney user.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://mymoney-jd0n.onrender.com",
      "description": "Production MyMoney API Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/budgets": {
      "get": {
        "summary": "List user budgets",
        "description": "Retrieve monthly budget limits set by the authenticated user for overall spending and specific categories.",
        "operationId": "listBudgets",
        "responses": {
          "200": {
            "description": "List of budget objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BudgetItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create or update a budget limit",
        "description": "Set or update a monthly budget limit for a category (or 'Overall') for the authenticated user.",
        "operationId": "createOrUpdateBudget",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created or updated budget item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key (mm_live_...)",
        "description": "Pass your MyMoney API Key generated from Settings (Authorization: Bearer mm_live_...)"
      }
    },
    "schemas": {
...
      "BudgetItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "example": "Food"
          },
          "amount": {
            "type": "number",
            "example": 5000.0
          }
        }
      },
      "BudgetCreate": {
        "type": "object",
        "required": [
          "category",
          "amount"
        ],
        "properties": {
          "category": {
            "type": "string",
            "description": "Category name or 'Overall'",
            "example": "Food"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01,
            "description": "Monthly budget limit amount",
            "example": 5000.0
          }
        }
      },
...
    }
  }
}